fix: correct dual-arm action indexing when gripper is stripped
- RelativeFrame.transform_action/transform_action_inv: use dynamic
  slice offset (7 for 14D, 6 for 12D) instead of hardcoded [7:13]
  to avoid index errors when GripperCloseEnv strips gripper dims
- GripperCloseEnv.step: guard intervene_action stripping with
  length check to handle already-12D actions
- DualFrankaEnv._get_observation: explicit dtype=np.float32 for
  gripper_position to match observation space definition

Made-with: Cursor

+++ rlinf/envs/realworld/franka/dual_franka_env.py
@@ -631,7 +631,7 @@ class DualFrankaEnv(gym.Env):
                 "gripper_position": np.array([
                     self._left_state.gripper_position,
                     self._right_state.gripper_position,
-                ]),
+                ], dtype=np.float32),
                 "tcp_force": np.concatenate([
                     self._left_state.tcp_force, self._right_state.tcp_force,
                 ]),
